Move `winit` module to its own file
